Compute JAVA_HOME properly if it is not set
In the case that JAVA_HOME was not set, and we are unable to
find /usr/libexec/java_home , we now compute the path correctly
based on the path to javac. Before this change, we were
incorrectly setting it to e.g. /usr/bin/javac , which made no sense
at all.

In addition, for clarity, we now always use JAVA_HOME to compute the
path of JAVAC, even if we used the path to `javac` to find JAVA_HOME.